% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{add_r_edges}
\alias{add_r_edges}
\title{Generate random edges in network}
\usage{
add_r_edges(start.net, r.rate, n.prob, r.trans.rate)
}
\arguments{
\item{start.net}{Network matrix to add random edges to}

\item{r.rate}{Rate at which individuals add random edges (similar to sociality)}

\item{n.prob}{individual probabilities of being involved in random edge generation}

\item{r.trans.rate}{Numeric indicating probability a random contact between a susceptible and infectious results in a new infection}
}
\value{
Updated contact matrix with random edges added
}
\description{
Function to generate random edges on top of network
}
