% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{next_state}
\alias{next_state}
\title{Next State}
\usage{
next_state(pre.status, age)
}
\arguments{
\item{pre.status}{Infection status that has expired}

\item{age}{age of individual for probability of moving to different states}
}
\value{
vector the next state
}
\description{
Take input infection state and age and return the next infection state and the time spent in that state
}
