% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{p_sevsymp}
\alias{p_sevsymp}
\title{Probability of severely symptomatic (will be hospitalized) infection}
\usage{
p_sevsymp(age)
}
\arguments{
\item{age}{age of person}
}
\value{
numeric of probability of having severe symptoms.
}
\description{
Age-dependent probabilities of having severe symptoms. Used in `next.state`
}
