% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{p_symp}
\alias{p_symp}
\title{Probability of symptomatic infection}
\usage{
p_symp(age)
}
\arguments{
\item{age}{age of person}
}
\value{
numeric of probability of clinical (symptomatic).
}
\description{
Age-dependent probabilities of having clinical (symptomatic) infection. Used in `next.state`. From https://www.nature.com/articles/s41591-020-0962-9/figures/8
}
