% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{sum.inf}
\alias{sum.inf}
\title{Summarize Infection}
\usage{
\method{sum}{inf}(x)
}
\arguments{
\item{x}{infections status vector from `ABM` run}
}
\value{
vector of length 9 with integer counts of number of agents in each class
}
\description{
Function to summarize infection vector with individual states into aggregate
}
