% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{t_latent}
\alias{t_latent}
\title{latent period}
\usage{
t_latent(n_E, shape.l = 4, scale.l = 1)
}
\arguments{
\item{n_E}{Number of draws}

\item{shape.l}{Shape of gamma distribution}

\item{scale.l}{Scale of gamma distribution}
}
\value{
numeric of time spent in latent period (E)
}
\description{
Draw from distribution of proposed latent periods. Used in `next.state`
}
