% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{t_msymp}
\alias{t_msymp}
\title{Mild infection duration}
\usage{
t_msymp(n_Im, shape.im = 6, scale.im = 1)
}
\arguments{
\item{n_Im}{number of draws}

\item{shape.im}{Shape parameter of gamma distn}

\item{scale.im}{Scale parameter of gamma distn}
}
\value{
numeric of time spent in mildly symptomatic (Im) state
}
\description{
Draw from distribution of proposed length spent with minor symptoms. Used in `next.state`
}
