% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{t_presymp}
\alias{t_presymp}
\title{Presymptomatic period}
\usage{
t_presymp(n_Ip, shape.ip = 10, scale.ip = 7)
}
\arguments{
\item{n_Ip}{Number of draws}

\item{shape}{Shape of gamma distribution}

\item{scale}{Scale of gamma distribution}
}
\value{
numeric of time spent in presymptomatic period (Ip)
}
\description{
Draw from distribution of proposed presymptomatic periods (basically incubation-latent periods) to model presymptomatic transmission. Used in `next.state`
}
