% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{t_sevsymp}
\alias{t_sevsymp}
\title{time spent severely symptomatic (in hospital)}
\usage{
t_sevsymp(n_Ih, shape.ih = 6, scale.ih = 0.5)
}
\arguments{
\item{n_Ih}{number of draws}

\item{shape.ih}{shape of gamma distn}

\item{scale.ih}{scale of gamma distn}
}
\value{
numeric of time spent hospitalized (Ih)
}
\description{
Draw from distribution of proposed time spent with severe symptoms before dying or being discharged. Used in `next.state`
}
