% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{t_til_nxt}
\alias{t_til_nxt}
\title{Time til Next state}
\usage{
t_til_nxt(pre.status)
}
\arguments{
\item{pre.status}{Infection status that has expired}
}
\value{
vector time to be spent in infection state. Time should be converted to `numeric` for subsequent use
}
\description{
Take input infection state the time spent in that state
}
