% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABM_funcs.R
\name{covid_abm}
\alias{covid_abm}
\title{COVID ABM}
\usage{
covid_abm(
  t.tot,
  dt,
  pop.ages,
  q.probs,
  n.probs,
  inf.mat,
  transition.mat,
  relation.mat
)
}
\arguments{
\item{t.tot}{total time to run simulation}

\item{dt}{time step of simulation to run}

\item{pop.ages}{vector corresponding to age of each individual}

\item{q.probs}{vector of individual probabilities of quarantining}

\item{n.probs}{vector of individual probabilities of being in random edge generation}

\item{inf.mat}{infection status matrix with column 1 filled with initial infection status, N rows, and t.tot/dt columns}

\item{transition.mat}{transition times matrix (time until moving into next state) with column 1 filled with initial times remaining in starting states, N rows, and t.tot/dt columns}

\item{relation.mat}{contact network array with dimensions NxN in which each ij entry represents relationship between agent i and agent j}
}
\value{
Matrix with columns corresponding to each infection state and rows filled with number people in each state at each time over which the simulation was run
}
\description{
Simulate COVID transmission across a network of N individuals
}
